consolidate types, cleanup unused/commented code
