+++ components/chat/sidebars/ImageSidebar.tsx
@@ -2,9 +2,9 @@ import { Tabs } from '@radix-ui/themes'
 
 import { ImageModelCard } from '@/components/cards/ImageModelCard'
 import { SlashCommandCard } from '@/components/cards/SlashCommandCard'
-import { EImageModel, EThread } from '@/convex/shared/types'
+import { EImageModel, EThread } from '@/convex/types'
 
-import type { TextToImageConfig } from '@/convex/shared/types'
+import type { TextToImageConfig } from '@/convex/types'
 
 export const ImageSidebar = ({ thread }: { thread: EThread; config: TextToImageConfig }) => {
   const model = thread.model?.type === 'image' ? (thread.model as EImageModel) : null
